+++ test/unit/origins/sdist/test_tarball.py
@@ -238,8 +238,8 @@ class TestTarball(SDistTestCase):
              mock.patch('builtins.open'):
             self.check_resolve(pkg)
         self.check_linkage(pkg, linkage={
-            'name': 'foo', 'type': 'system', 'auto_link': False,
-            'pcnames': ['foo'], 'pkg_config_path': [self.pkgconfdir(None)],
+            'name': 'foo', 'type': 'system', 'pcnames': ['foo'],
+            'pkg_config_path': [self.pkgconfdir(None)],
         })
 
     def test_infer_build_override(self):
@@ -319,8 +319,8 @@ class TestTarball(SDistTestCase):
         self.assertEqual(pkg.builders, [builder])
         self.check_resolve(pkg)
         self.check_linkage(pkg, linkage={
-            'name': 'foo', 'type': 'path', 'auto_link': False,
-            'pcnames': ['foo'], 'pkg_config_path': [self.pkgconfdir(None)],
+            'name': 'foo', 'type': 'path', 'pcnames': ['foo'],
+            'pkg_config_path': [self.pkgconfdir(None)],
         })
 
         with mock.patch('subprocess.run') as mrun: