+++ test/unit/origins/test_apt.py
@@ -53,8 +53,8 @@ class TestApt(OriginTest):
                         for i in iterate(submodules))
 
             linkage = {
-                'name': depname, 'type': 'system', 'auto_link': False,
-                'pcnames': [depname], 'pkg_config_path': [self.pkgconfdir],
+                'name': depname, 'type': 'system', 'pcnames': [depname],
+                'pkg_config_path': [self.pkgconfdir],
             }
 
         with mock.patch('subprocess.run', mock_run), \
@@ -186,8 +186,8 @@ class TestApt(OriginTest):
         )
         self.check_resolve_all([pkg], ['libfoo-dev'])
         self.check_linkage(pkg, submodules=['sub'], linkage={
-            'name': 'foo[sub]', 'type': 'system', 'auto_link': False,
-            'pcnames': ['foo[sub]'], 'pkg_config_path': [self.pkgconfdir],
+            'name': 'foo[sub]', 'type': 'system', 'pcnames': ['foo[sub]'],
+            'pkg_config_path': [self.pkgconfdir],
         })
 
         pkg = self.make_package('foo', submodules=submodules_optional)
@@ -200,8 +200,8 @@ class TestApt(OriginTest):
         )
         self.check_resolve_all([pkg], ['libfoo-dev'])
         self.check_linkage(pkg, submodules=['sub'], linkage={
-            'name': 'foo[sub]', 'type': 'system', 'auto_link': False,
-            'pcnames': ['foo[sub]'], 'pkg_config_path': [self.pkgconfdir],
+            'name': 'foo[sub]', 'type': 'system', 'pcnames': ['foo[sub]'],
+            'pkg_config_path': [self.pkgconfdir],
         })
 
     def test_invalid_submodule(self):